requirements.txt
albumentations
annotated-types
bypy
certifi
charset-normalizer
contourpy
cycler
Cython
dill
easydict
einops
einsum
elasticdeform
facenet-pytorch
filelock
fonttools
fsspec
ftfy
gdown
huggingface-hub
imageio
imgaug
insightface
isort
joblib
kiwisolver
lazy_loader
matplotlib
mkl-service
multiprocess
networkx
numpy
onnx
opencv-python
opencv-python-headless
pandas
prettytable
protobuf
pydantic
pydantic_core
pyparsing
PySocks
pytz
PyWavelets
regex
requests
requests-toolbelt
safetensors
scikit-image
scikit-learn
scipy
seaborn
shapely
threadpoolctl
tifffile
timm
tomli
tqdm
tzdata
urllib3
wcwidth
webencodings
yapf
